Round månedsbeløp for ytelser to nearest krone

+++ client/src/test/kotlin/no/nav/su/se/bakover/client/oppdrag/avstemming/AvstemmingDataBuilderTest.kt
@@ -40,7 +40,7 @@ internal class AvstemmingDataBuilderTest {
             ),
             total = AvstemmingDataRequest.Totaldata(
                 totalAntall = 5,
-                totalBelop = 18000.0.toBigDecimal(),
+                totalBelop = 18000.toBigDecimal(),
                 fortegn = AvstemmingDataRequest.Fortegn.TILLEGG
             ),
             periode = AvstemmingDataRequest.Periodedata(
@@ -49,16 +49,16 @@ internal class AvstemmingDataBuilderTest {
             ),
             grunnlag = AvstemmingDataRequest.Grunnlagdata(
                 godkjentAntall = 2,
-                godkjentBelop = 1600.0.toBigDecimal(),
+                godkjentBelop = 1600.toBigDecimal(),
                 godkjentFortegn = AvstemmingDataRequest.Fortegn.TILLEGG,
                 varselAntall = 1,
-                varselBelop = 1400.0.toBigDecimal(),
+                varselBelop = 1400.toBigDecimal(),
                 varselFortegn = AvstemmingDataRequest.Fortegn.TILLEGG,
                 avvistAntall = 1,
-                avvistBelop = 10000.0.toBigDecimal(),
+                avvistBelop = 10000.toBigDecimal(),
                 avvistFortegn = AvstemmingDataRequest.Fortegn.TILLEGG,
                 manglerAntall = 1,
-                manglerBelop = 5000.0.toBigDecimal(),
+                manglerBelop = 5000.toBigDecimal(),
                 manglerFortegn = AvstemmingDataRequest.Fortegn.TILLEGG
             ),
             detalj = listOf(
@@ -97,7 +97,7 @@ internal class AvstemmingDataBuilderTest {
 }
 
 private val zoneId = ZoneId.of("Europe/Oslo")
-fun lagUtbetalingLinje(fraOgMed: LocalDate, tilOgMed: LocalDate, beløp: Double) = Utbetalingslinje(
+fun lagUtbetalingLinje(fraOgMed: LocalDate, tilOgMed: LocalDate, beløp: Int) = Utbetalingslinje(
     id = UUID30.randomUUID(),
     opprettet = fraOgMed.atStartOfDay(zoneId).toTidspunkt(),
     fraOgMed = fraOgMed,
@@ -154,7 +154,7 @@ private val simulering = Simulering(
     gjelderId = fnr,
     gjelderNavn = "",
     datoBeregnet = idag(),
-    nettoBeløp = 0.0,
+    nettoBeløp = 0,
     periodeList = listOf()
 )
 fun alleUtbetalinger() = listOf(
@@ -163,8 +163,8 @@ fun alleUtbetalinger() = listOf(
         opprettet = 1.mars(2020),
         status = Kvittering.Utbetalingsstatus.OK,
         linjer = listOf(
-            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 100.0),
-            lagUtbetalingLinje(1.april(2020), 30.april(2020), 200.0)
+            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 100),
+            lagUtbetalingLinje(1.april(2020), 30.april(2020), 200)
         )
     ),
     lagUtbetaling(
@@ -172,8 +172,8 @@ fun alleUtbetalinger() = listOf(
         opprettet = 1.mars(2020),
         status = Kvittering.Utbetalingsstatus.OK,
         linjer = listOf(
-            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 600.0),
-            lagUtbetalingLinje(1.april(2020), 30.april(2020), 700.0)
+            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 600),
+            lagUtbetalingLinje(1.april(2020), 30.april(2020), 700)
         )
     ),
     lagUtbetaling(
@@ -181,9 +181,9 @@ fun alleUtbetalinger() = listOf(
         opprettet = 2.mars(2020),
         status = Kvittering.Utbetalingsstatus.OK_MED_VARSEL,
         linjer = listOf(
-            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 400.0),
-            lagUtbetalingLinje(1.april(2020), 30.april(2020), 500.0),
-            lagUtbetalingLinje(1.mai(2020), 31.mai(2020), 500.0)
+            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 400),
+            lagUtbetalingLinje(1.april(2020), 30.april(2020), 500),
+            lagUtbetalingLinje(1.mai(2020), 31.mai(2020), 500)
         )
     ),
     lagUtbetaling(
@@ -191,10 +191,10 @@ fun alleUtbetalinger() = listOf(
         opprettet = 1.mars(2020),
         status = Kvittering.Utbetalingsstatus.FEIL,
         linjer = listOf(
-            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 1000.0),
-            lagUtbetalingLinje(1.april(2020), 30.april(2020), 2000.0),
-            lagUtbetalingLinje(1.mai(2020), 31.mai(2020), 3000.0),
-            lagUtbetalingLinje(1.juni(2020), 30.juni(2020), 4000.0)
+            lagUtbetalingLinje(1.mars(2020), 31.mars(2020), 1000),
+            lagUtbetalingLinje(1.april(2020), 30.april(2020), 2000),
+            lagUtbetalingLinje(1.mai(2020), 31.mai(2020), 3000),
+            lagUtbetalingLinje(1.juni(2020), 30.juni(2020), 4000)
         )
     ),
     lagUtbetaling(
@@ -202,7 +202,7 @@ fun alleUtbetalinger() = listOf(
         opprettet = 2.mars(2020),
         status = null,
         linjer = listOf(
-            lagUtbetalingLinje(1.januar(2020), 31.desember(2020), 5000.0)
+            lagUtbetalingLinje(1.januar(2020), 31.desember(2020), 5000)
         )
     )
 )
