+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/beregning/Månedsberegning.kt
@@ -3,7 +3,7 @@ package no.nav.su.se.bakover.domain.beregning
 import no.nav.su.se.bakover.common.periode.PeriodisertInformasjon
 
 interface Månedsberegning : PeriodisertInformasjon {
-    fun getSumYtelse(): Double
+    fun getSumYtelse(): Int
     fun getSumFradrag(): Double
     fun getBenyttetGrunnbeløp(): Int
     fun getSats(): Sats
